Fix finding BSS sections (#624)
* Fix finding BSS sections

There can be multiple NOBITS sections in an ELF binary - and we were taking
the first one as being the BSS section. In certain cases this doesn't work
(including the latest python 3.12 binary in conda) - since there can
be '.tbss' or '.sbss' sections instead.

Fix by performing an additional filter step using the section name.

* fix windows build with newer indicatif

Granulate backport notes: We only cherry-picked the changes to code and not dependencies, as the current versions of packages are enough to use the new code.